Add a "stack" callback
This adds a callback, like the settings, zip, and post callbacks, which
runs between setting up the CloudFormation stack and deploying it. This
can be used to enable projects to extend the stack, as proposed
[here](https://github.com/Miserlou/Zappa/issues/399). This could allow
automated provisioning of Route53 entries, or anything else that can
be provisioned with CloudFormation templates that should be deployed
along side a Zappa application.